chore: change banner format
