+++ package.json
@@ -10,10 +10,8 @@
   },
   "dependencies": {
     "@rails/actioncable": "^7.0.4",
-    "@react-three/fiber": "^8.11.5",
     "font-awesome": "^4.7.0",
     "react": "^18.2.0",
-    "react-3d-icons": "^1.0.2",
     "react-cssfx-loading": "^2.1.0",
     "react-dom": "^18.2.0",
     "react-ga": "^3.3.1",
@@ -24,7 +22,6 @@
     "react-router-dom": "^6.4.2",
     "react-use-websocket": "^3.0.0",
     "remark-gfm": "^3.0.1",
-    "three": "^0.150.1",
     "typed.js": "^2.0.12"
   },
   "devDependencies": {
@@ -37,5 +34,10 @@
     "prettier": "^2.7.1",
     "tailwindcss": "^3.2.4",
     "vite": "^3.2.0"
-  }
+  },
+  "description": "<img src=\"https://camo.githubusercontent.com/4e4a3b5c3e9c00501ec866e2f2466c5a6032f838aca5f2cf3b14450e39e8a2f0/68747470733a2f2f696d672e736869656c64732e696f2f62616467652f72656163742532302d2532333230323332612e7376673f267374796c653d666f722d7468652d6261646765266c6f676f3d7265616374266c6f676f436f6c6f723d253233363144414642\"/>",
+  "main": "vite.config.js",
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
