+++ src/components/Banner.jsx
@@ -15,16 +15,13 @@ const Banner = () => {
   //   });
   // });
   return (
-    <div className="flex flex-col justify-center text-center gap-y-4" id="home">
-      <div className="font-primary text-2xl text-center font-extrabold mt-40 w-[80%] sm:w-[100%] mx-auto">
-        Scan your project &nbsp;
-        <span className="text-transparent bg-clip-text bg-gradient-to-r from-purple-400 to-pink-600">
-          for better security
-        </span>
-      </div>
-      <span className="font-secondary">
-        This app allow to check vulnerabilities of Github projects
-      </span>
+    <div className="text-center">
+      <h1>
+        Make your project safer with <strong>SCANREPO</strong>
+      </h1>
+      <h2 className="text-[12px] text-gray-500">
+        by findind vulnerabilities directly into your github repositories
+      </h2>
     </div>
   );
 };
