+++ src/components/Footer.jsx
@@ -4,7 +4,7 @@ const Footer = () => {
   return (
     <div className="sticky bottom-0 flex flex-col items-center h-auto bg-black-github gap-y-1 px-7 py-10 text-sm md:text-md lg:text-xl">
       <p>The best way to secure your project at start</p>
-      <span className="text-xs">© 2022 Scanrepo, Inc.</span>
+      <span className="text-xs">© 2022 Scanrepo, Francois Robino</span>
     </div>
   );
 };
