+++ src/components/Header.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import SocialNetwork from "./SocialNetwork";
 
-const Logo = () => {
+const Header = () => {
   return (
     <div className="flex flex-row items-center justify-between md:items-start p-6 top-0">
       <div className="logo-area flex flex-row items-center gap-x-1">
@@ -15,4 +15,4 @@ const Logo = () => {
   );
 };
 
-export default Logo;
+export default Header;
