--- src/components/InputField.jsx
@@ -1,69 +1,0 @@
-import React, { useState } from "react";
-import { useForm } from "react-hook-form";
-import { apiBaseUrl } from "../utils/config";
-import { useNavigate } from "react-router-dom";
-
-const Input = () => {
-  const { register, handleSubmit } = useForm();
-  const navigate = useNavigate();
-  const [state, setState] = useState("");
-
-  const onSubmit = async (data) => {
-    if (state.split("//")[0] === "https:") {
-      try {
-        const response = await fetch(`${apiBaseUrl}/projects`, {
-          method: "POST",
-          credentials: "include",
-          headers: {
-            "Content-type": "application/json",
-          },
-          body: JSON.stringify({ project: data }),
-        });
-
-        const project = await response.json();
-        if (!response.ok) {
-          return;
-        }
-        navigate(`/project/${project.uuid}`);
-      } catch {}
-    } else {
-      navigate(`/repos/${state}`);
-    }
-  };
-
-  return (
-    <form
-      onSubmit={handleSubmit(onSubmit)}
-      onChange={(e) => setState(e.target.value)}
-      className="pr-0 md:pr-20 fle"
-    >
-      <div className="gap-y-3">
-        <span className="font-primary text-transparent bg-clip-text bg-gradient-to-r from-purple-400 to-pink-600 pb-5">
-          Github username or url of project
-        </span>
-        <div>
-          <div className="sm:flex-row text-center md:text-left lg:gap-x-5 md:gap-x-5 gap-y-5 mt-5">
-            <input
-              type="text"
-              placeholder="Francois973 or https://github.com/francois973/vulnerability-checker-back-end"
-              className="border-solid border-2 border-blue-light bg-dark-color p-2 rounded-md transition ease-in-out delay-50 hover:border-pink-600 focus:outline-none h-[40px] mr-2 w-[80%]"
-              {...register("username")}
-              {...register("url")}
-              required
-            />
-            <input
-              type="submit"
-              value="GO"
-              className="py-0 cursor-pointer bg-gradient-to-r from-purple-400 to-pink-600 font-bold p-2 rounded-md hover:opacity-60 transition ease-in-out delay-50 h-[40px] mb-2 w-[15%]"
-            />
-          </div>
-        </div>
-        <i className="font-secondary p-0 text-sm text-gray-500 md:text-center pt-2 sm:text-left">
-          Your informations will not be stored under any circumstances
-        </i>
-      </div>
-    </form>
-  );
-};
-
-export default Input;
