+++ src/components/Layout.jsx
@@ -1,10 +1,11 @@
 import React from "react";
 import Footer from "./Footer";
-import Logo from "./Logo";
+import Header from "./Header";
 
 const Layout = ({ children }) => {
   return (
     <div className="flex flex-col min-h-screen justify-between bg-black-github text-white">
+      <Header />
       <main className="container mx-auto mb-auto">{children}</main>
       <Footer />
     </div>
