+++ src/components/SocialNetwork.jsx
@@ -1,33 +1,7 @@
 import React from "react";
-import { github, linkedin } from "react-3d-icons";
-import { Icon } from "react-3d-icons";
 
 const SocialNetwork = () => {
-  return (
-    <div className="flex flex-row">
-      <div style={{ height: "50px", width: "50px" }}>
-        <a href="https://github.com/francois973/" target="_blank">
-          {" "}
-          <Icon
-            file={github}
-            color={"white"}
-            scale={10}
-            className="hover:scale-110 transition-all ease-out duration-300"
-          />
-        </a>
-      </div>
-      <div style={{ height: "50px", width: "50px" }}>
-        <a href="https://www.linkedin.com/in/francoisrobino/" target="_blank">
-          <Icon
-            file={linkedin}
-            color={"white"}
-            scale={10}
-            className="hover:scale-110 transition-all ease-out duration-300"
-          />
-        </a>
-      </div>
-    </div>
-  );
+  return <div className="flex flex-row"></div>;
 };
 
 export default SocialNetwork;
