+++ src/components/Vulnerabilities.jsx
@@ -1,10 +1,12 @@
-import React from "react";
+import React, { useEffect } from "react";
 import Vulnerability from "../components/Vulnerability";
 
 const Vulnerabilities = ({ vulnerabilities, selectedEcosystem }) => {
   const filteredVulnerabilities = selectedEcosystem
     ? vulnerabilities.filter((v) => v.ecosystem === selectedEcosystem)
     : vulnerabilities;
+
+  const active = filteredVulnerabilities;
 
   return (
     <div>
@@ -13,7 +15,10 @@ const Vulnerabilities = ({ vulnerabilities, selectedEcosystem }) => {
           <p>No vulnerabilities here !</p>
         ) : (
           filteredVulnerabilities.map((v, i) => (
-            <li key={i} className="h-auto mx-3 md:w-[600px] w-[80vw]">
+            <li
+              key={i}
+              className="h-auto mx-3 md:w-[600px] w-[80vw] transition ease-out duration-500"
+            >
               <Vulnerability vulnerability={v} />
             </li>
           ))
