+++ src/components/Vulnerability.jsx
@@ -4,7 +4,7 @@ import Severity from "./Severity";
 
 const Vulnerability = ({ vulnerability }) => {
   return (
-    <div className="p-6  bg-dark-color rounded-md sm:h-[40.685678073510772vh] h-auto flex flex-col justify-between gap-y-2">
+    <div className="p-6 bg-dark-color rounded-md sm:h-[40.685678073510772vh] h-auto flex flex-col justify-between gap-y-2">
       <div className="flex flex-row justify-between">
         <strong>{vulnerability.stack}</strong>
         {vulnerability.ecosystem === "RUBYGEMS" ? (
