+++ src/pages/Home.jsx
@@ -2,9 +2,7 @@ import { React, useEffect } from "react";
 import Banner from "../components/Banner.jsx";
 import ReactGA from "react-ga4";
 import Available from "../components/Available.jsx";
-import InputField from "../components/InputField";
-import Logo from "../components/Logo.jsx";
-import Footer from "../components/Footer.jsx";
+import InputField from "../components/StartForm";
 
 const Home = () => {
   useEffect(() => {
@@ -13,15 +11,9 @@ const Home = () => {
 
   return (
     <div className="md:h-[80vh] h-auto overflow-y-hidden">
-      <Logo />
-      <div className="flex flex-col gap-y-[6rem] items-center justify-center">
-        <Banner />
-        <div className="flex flex-col md:flex-row justify-center items-start sm:items-center ">
-          <InputField />
-          <div className="md:h-80 md:w-0 border-2 w-80 h-0 m-auto"></div>
-          <Available />
-        </div>
-      </div>
+      <Banner />
+      <InputField />
+      <Available />
     </div>
   );
 };
