+++ src/pages/Project.jsx
@@ -7,7 +7,6 @@ import Vulnerabilities from "../components/Vulnerabilities";
 import { apiBaseUrl } from "../utils/config";
 import LoaderUsername from "../components/LoaderUsername";
 import ReactGA from "react-ga4";
-import Logo from "../components/Logo";
 
 function Project() {
   const [project, setProject] = useState();
@@ -51,7 +50,6 @@ function Project() {
   }
   return (
     <div>
-      <Logo />
       <div className="ml-7 mt-20">
         <h2 className="text-lg mb-2 ">
           <strong>Repository:</strong>{" "}
