+++ src/pages/Repositories.jsx
@@ -5,7 +5,6 @@ import LanguagesRepos from "../components/LanguagesRepos";
 import LoaderUsername from "../components/LoaderUsername";
 import { apiBaseUrl } from "../utils/config";
 import ReactGA from "react-ga4";
-import Logo from "../components/Logo";
 
 const Repositories = () => {
   const { username } = useParams();
@@ -44,7 +43,6 @@ const Repositories = () => {
 
   return (
     <>
-      <Logo />
       <div className="relative">
         {isLoading
           ? null
