+++ tailwind.config.cjs
@@ -4,7 +4,7 @@ module.exports = {
   theme: {
     extend: {
       colors: {
-        "black-github": "#010409",
+        "black-github": "#22272e",
         "blue-light": "#CDDDDD",
         "dark-color": "#051014",
         "green-color": "#238636",
@@ -18,9 +18,7 @@ module.exports = {
       },
     },
     fontFamily: {
-      primary: "Orbitron",
-      secondary: "Rajdhani",
-      tertiary: "Aldrich",
+      font: "Roboto",
     },
   },
   plugins: [],
